Move classes that have nothing to do with GraalVM outside of graal/
